pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

# =============================================================================
# CNameCheck — Embedded C naming-convention linter
# =============================================================================

[project]
name            = "cnamecheck"
version         = "1.0.0"
description     = "Embedded C naming-convention linter (Barr-C / MISRA-C complementary)"
readme          = "README.md"
license         = { text = "MIT" }
requires-python = ">=3.10"
dependencies    = ["pyyaml>=6.0,<7.0"]
keywords        = ["c", "linter", "naming-convention", "barr-c", "misra", "embedded"]

[project.urls]
Homepage      = "https://github.com/dermot-murphy/CNameCheck"
Repository    = "https://github.com/dermot-murphy/CNameCheck"
"Bug Tracker" = "https://github.com/dermot-murphy/CNameCheck/issues"

# `cnamecheck` is available as a command after `pip install .` or `pipx install .`
[project.scripts]
cnamecheck = "cnamecheck:main"

# =============================================================================
# Package layout
#
# cnamecheck.py and _version.py live in src/ and are installed as top-level
# Python modules.  The three dictionary files are installed alongside them in
# {sys.prefix}/share/cnamecheck/ so that the tool can locate them at runtime
# regardless of whether it was run from a source checkout or a pip install.
# =============================================================================

[tool.setuptools]
py-modules = ["cnamecheck", "_version"]

[tool.setuptools.package-dir]
"" = "src"

# data_files installs the bundled dictionaries to {sys.prefix}/share/cnamecheck/
# cnamecheck._data_file() checks that directory as a fallback when the files
# are not found alongside __file__ (which is the case for regular pip installs).
[tool.setuptools.data-files]
"share/cnamecheck" = [
    "src/c_keywords.txt",
    "src/c_stdlib_names.txt",
    "src/c_spell_dict.txt",
]

# =============================================================================
# Tool configuration
# =============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
